+++ database/migrations/2022_06_21_233735_create_voucer_table.php
@@ -16,7 +16,7 @@ class CreateVoucerTable extends Migration
         Schema::create('voucher', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('code');
-            $table->double('discount');
+            $table->decimal('discount', 2, 2);
             $table->integer('quota');
             $table->integer('used_qouta')->default(0);
             $table->timestamps();
